// src/grocyClient/model/batteryDetailsResponse.ts
/**
 * grocy REST API
 * Authentication is done via API keys (header *GROCY-API-KEY* or same named query parameter), which you can manage [here](https://demo.grocy.info/manageapikeys).<br>Additionally requests from within the frontend are also valid (via session cookie).
 *
 * The version of the OpenAPI document: 2.6.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "../api";
import { Battery } from "./battery";

export class BatteryDetailsResponse {
  "chore"?: Battery;
  /**
   * When this battery was last charged
   */
  "lastCharged"?: Date;
  /**
   * How often this battery was charged so far
   */
  "chargeCyclesCount"?: number;
  "nextEstimatedChargeTime"?: Date;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "chore",
      baseName: "chore",
      type: "Battery"
    },
    {
      name: "lastCharged",
      baseName: "last_charged",
      type: "Date"
    },
    {
      name: "chargeCyclesCount",
      baseName: "charge_cycles_count",
      type: "number"
    },
    {
      name: "nextEstimatedChargeTime",
      baseName: "next_estimated_charge_time",
      type: "Date"
    }
  ];

  static getAttributeTypeMap() {
    return BatteryDetailsResponse.attributeTypeMap;
  }
}