// src/grocyClient/model/currentBatteryResponse.ts
/**
 * grocy REST API
 * Authentication is done via API keys (header *GROCY-API-KEY* or same named query parameter), which you can manage [here](https://demo.grocy.info/manageapikeys).<br>Additionally requests from within the frontend are also valid (via session cookie).
 *
 * The version of the OpenAPI document: 2.6.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "../api";

export class CurrentBatteryResponse {
  "batteryId"?: number;
  "lastTrackedTime"?: Date;
  /**
   * The next estimated charge time of this battery, 2999-12-31 23:59:59 when the given battery has no charge_interval_days defined
   */
  "nextEstimatedChargeTime"?: Date;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "batteryId",
      baseName: "battery_id",
      type: "number"
    },
    {
      name: "lastTrackedTime",
      baseName: "last_tracked_time",
      type: "Date"
    },
    {
      name: "nextEstimatedChargeTime",
      baseName: "next_estimated_charge_time",
      type: "Date"
    }
  ];

  static getAttributeTypeMap() {
    return CurrentBatteryResponse.attributeTypeMap;
  }
}