// src/grocyClient/model/currentChoreResponse.ts
/**
 * grocy REST API
 * Authentication is done via API keys (header *GROCY-API-KEY* or same named query parameter), which you can manage [here](https://demo.grocy.info/manageapikeys).<br>Additionally requests from within the frontend are also valid (via session cookie).
 *
 * The version of the OpenAPI document: 2.6.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "../api";

export class CurrentChoreResponse {
  "choreId"?: number;
  "lastTrackedTime"?: Date;
  /**
   * The next estimated execution time of this chore, 2999-12-31 23:59:59 when the given chore has a period_type of manually
   */
  "nextEstimatedExecutionTime"?: Date;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "choreId",
      baseName: "chore_id",
      type: "number"
    },
    {
      name: "lastTrackedTime",
      baseName: "last_tracked_time",
      type: "Date"
    },
    {
      name: "nextEstimatedExecutionTime",
      baseName: "next_estimated_execution_time",
      type: "Date"
    }
  ];

  static getAttributeTypeMap() {
    return CurrentChoreResponse.attributeTypeMap;
  }
}