// src/grocyClient/model/currentStockResponse.ts
/**
 * grocy REST API
 * Authentication is done via API keys (header *GROCY-API-KEY* or same named query parameter), which you can manage [here](https://demo.grocy.info/manageapikeys).<br>Additionally requests from within the frontend are also valid (via session cookie).
 *
 * The version of the OpenAPI document: 2.6.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "../api";
import { Product } from "./product";

export class CurrentStockResponse {
  "productId"?: number;
  "amount"?: number;
  "amountAggregated"?: number;
  "amountOpened"?: number;
  "amountOpenedAggregated"?: number;
  /**
   * The next best before date for this product
   */
  "bestBeforeDate"?: string;
  /**
   * Indicates wheter this product has sub-products or not / if the fields `amount_aggregated` and `amount_opened_aggregated` are filled
   */
  "isAggregatedAmount"?: boolean;
  "product"?: Product;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "productId",
      baseName: "product_id",
      type: "number"
    },
    {
      name: "amount",
      baseName: "amount",
      type: "number"
    },
    {
      name: "amountAggregated",
      baseName: "amount_aggregated",
      type: "number"
    },
    {
      name: "amountOpened",
      baseName: "amount_opened",
      type: "number"
    },
    {
      name: "amountOpenedAggregated",
      baseName: "amount_opened_aggregated",
      type: "number"
    },
    {
      name: "bestBeforeDate",
      baseName: "best_before_date",
      type: "string"
    },
    {
      name: "isAggregatedAmount",
      baseName: "is_aggregated_amount",
      type: "boolean"
    },
    {
      name: "product",
      baseName: "product",
      type: "Product"
    }
  ];

  static getAttributeTypeMap() {
    return CurrentStockResponse.attributeTypeMap;
  }
}