// src/grocyClient/model/currentVolatilStockResponse.ts
/**
 * grocy REST API
 * Authentication is done via API keys (header *GROCY-API-KEY* or same named query parameter), which you can manage [here](https://demo.grocy.info/manageapikeys).<br>Additionally requests from within the frontend are also valid (via session cookie).
 *
 * The version of the OpenAPI document: 2.6.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "../api";
import { CurrentStockResponse } from "./currentStockResponse";

export class CurrentVolatilStockResponse {
  "expiringProducts"?: Array<CurrentStockResponse>;
  "expiredProducts"?: Array<CurrentStockResponse>;
  "missingProducts"?: Array<object>;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "expiringProducts",
      baseName: "expiring_products",
      type: "Array<CurrentStockResponse>"
    },
    {
      name: "expiredProducts",
      baseName: "expired_products",
      type: "Array<CurrentStockResponse>"
    },
    {
      name: "missingProducts",
      baseName: "missing_products",
      type: "Array<object>"
    }
  ];

  static getAttributeTypeMap() {
    return CurrentVolatilStockResponse.attributeTypeMap;
  }
}