// src/grocyClient/model/inlineObject1.ts
/**
 * grocy REST API
 * Authentication is done via API keys (header *GROCY-API-KEY* or same named query parameter), which you can manage [here](https://demo.grocy.info/manageapikeys).<br>Additionally requests from within the frontend are also valid (via session cookie).
 *
 * The version of the OpenAPI document: 2.6.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "../api";
import { StockTransactionType } from "./stockTransactionType";

export class InlineObject1 {
  /**
   * The amount to add - please note that when tare weight handling for the product is enabled, this needs to be the amount including the container weight (gross), the amount to be posted will be automatically calculated based on what is in stock and the defined tare weight
   */
  "amount"?: number;
  /**
   * The best before date of the product to add, when omitted, the current date is used
   */
  "bestBeforeDate"?: string;
  "transactionType"?: StockTransactionType;
  /**
   * The price per purchase quantity unit in configured currency
   */
  "price"?: number;
  /**
   * If omitted, the default location of the product is used
   */
  "locationId"?: number;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "amount",
      baseName: "amount",
      type: "number"
    },
    {
      name: "bestBeforeDate",
      baseName: "best_before_date",
      type: "string"
    },
    {
      name: "transactionType",
      baseName: "transaction_type",
      type: "StockTransactionType"
    },
    {
      name: "price",
      baseName: "price",
      type: "number"
    },
    {
      name: "locationId",
      baseName: "location_id",
      type: "number"
    }
  ];

  static getAttributeTypeMap() {
    return InlineObject1.attributeTypeMap;
  }
}